Fixed Steven's broken code.
